Zod for Edit and Checkout (best practice: use unknown type first and then validate the coming data in server action)

+++ petsoft/src/actions/actions.ts
@@ -2,7 +2,7 @@
 
 import { prisma } from '@/lib/prisma'
 import { sleep } from '@/lib/utils'
-import { petFormSchema } from '@/lib/validations'
+import { petFormSchema, petIdSchema } from '@/lib/validations'
 import { revalidatePath } from 'next/cache'
 
 export async function addPet(pet: unknown) {
@@ -31,12 +31,22 @@ export async function addPet(pet: unknown) {
 
 export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000)
+
+  const validatedPetId = petIdSchema.safeParse(petId)
+  const validatedPet = petFormSchema.safeParse(newPetData)
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: 'invalid pet data.',
+    }
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     })
   } catch (error) {
     return {
@@ -49,10 +59,17 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function checkoutPet(petId: unknown) {
   await sleep(1000)
 
+  const validatedPetId = petIdSchema.safeParse(petId)
+
+  if (!validatedPetId.success) {
+    return {
+      message: 'invalid pet data.',
+    }
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     })
   } catch (error) {
